requirements-dev.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements-dev.in, then run this to update:
#
#    pip-compile requirements-dev.in
#
-e .
alabaster==0.7.7          # via sphinx
babel==2.2.0              # via sphinx
beautifulsoup4==4.4.1     # via webtest
click==6.3                # via pip-tools
django-filter==0.12.0     # via django-viewflow
django-fsm==2.3.0         # via django-viewflow
django-viewflow==0.9.0
django-webtest==1.7.8     # via django-viewflow
Django==1.9.4             # via django-viewflow
docutils==0.12            # via sphinx
first==2.0.1              # via pip-tools
flake8==2.5.4
isort==4.2.2
Jinja2==2.8               # via sphinx
MarkupSafe==0.23          # via jinja2
mccabe==0.4.0
mock==1.3.0               # via django-viewflow
pbr==1.8.1                # via mock
pep257==0.7.0
pep8-naming==0.3.3
pep8==1.7.0               # via flake8
pip-tools==1.6
pluggy==0.3.1             # via tox
py==1.4.31                # via pytest, tox
pyenchant==1.6.6
pyflakes==1.0.0           # via flake8
Pygments==2.1.3           # via sphinx
pytest-django==2.9.1
pytest==2.9.0
pytz==2015.7              # via babel
singledispatch==3.4.0.3   # via django-viewflow
six==1.10.0               # via mock, pip-tools, singledispatch, sphinx, sphinxcontrib-spelling, webtest
snowballstemmer==1.2.1    # via sphinx
sphinx-rtd-theme==0.1.9   # via sphinx
sphinx==1.3.6
sphinxcontrib-spelling==2.1.2
tox==2.3.1
virtualenv==15.0.0        # via tox
waitress==0.8.10          # via webtest
WebOb==1.5.1              # via webtest
webtest==2.0.20           # via django-viewflow, django-webtest
